Cleanup related to context refactor
