+++ app/app.go
@@ -14,6 +14,7 @@ import (
 	"github.com/mattermost/mattermost-server/mlog"
 	"github.com/mattermost/mattermost-server/model"
 	"github.com/mattermost/mattermost-server/services/httpservice"
+	"github.com/mattermost/mattermost-server/services/timezones"
 	"github.com/mattermost/mattermost-server/utils"
 	goi18n "github.com/nicksnyder/go-i18n/i18n"
 )
@@ -42,6 +43,7 @@ type App struct {
 	Saml             einterfaces.SamlInterface
 
 	HTTPService httpservice.HTTPService
+	Timezones   *timezones.Timezones
 }
 
 func New(options ...AppOption) *App {
@@ -130,57 +132,6 @@ func (a *App) Handle404(w http.ResponseWriter, r *http.Request) {
 	mlog.Debug(fmt.Sprintf("%v: code=404 ip=%v", r.URL.Path, utils.GetIpAddress(r)))
 
 	utils.RenderWebAppError(a.Config(), w, r, err, a.AsymmetricSigningKey())
-}
-
-func (a *App) StartElasticsearch() {
-	a.Srv.Go(func() {
-		if err := a.Elasticsearch.Start(); err != nil {
-			mlog.Error(err.Error())
-		}
-	})
-
-	a.AddConfigListener(func(oldConfig *model.Config, newConfig *model.Config) {
-		if !*oldConfig.ElasticsearchSettings.EnableIndexing && *newConfig.ElasticsearchSettings.EnableIndexing {
-			a.Srv.Go(func() {
-				if err := a.Elasticsearch.Start(); err != nil {
-					mlog.Error(err.Error())
-				}
-			})
-		} else if *oldConfig.ElasticsearchSettings.EnableIndexing && !*newConfig.ElasticsearchSettings.EnableIndexing {
-			a.Srv.Go(func() {
-				if err := a.Elasticsearch.Stop(); err != nil {
-					mlog.Error(err.Error())
-				}
-			})
-		} else if *oldConfig.ElasticsearchSettings.Password != *newConfig.ElasticsearchSettings.Password || *oldConfig.ElasticsearchSettings.Username != *newConfig.ElasticsearchSettings.Username || *oldConfig.ElasticsearchSettings.ConnectionUrl != *newConfig.ElasticsearchSettings.ConnectionUrl || *oldConfig.ElasticsearchSettings.Sniff != *newConfig.ElasticsearchSettings.Sniff {
-			a.Srv.Go(func() {
-				if *oldConfig.ElasticsearchSettings.EnableIndexing {
-					if err := a.Elasticsearch.Stop(); err != nil {
-						mlog.Error(err.Error())
-					}
-					if err := a.Elasticsearch.Start(); err != nil {
-						mlog.Error(err.Error())
-					}
-				}
-			})
-		}
-	})
-
-	a.AddLicenseListener(func() {
-		if a.License() != nil {
-			a.Srv.Go(func() {
-				if err := a.Elasticsearch.Start(); err != nil {
-					mlog.Error(err.Error())
-				}
-			})
-		} else {
-			a.Srv.Go(func() {
-				if err := a.Elasticsearch.Stop(); err != nil {
-					mlog.Error(err.Error())
-				}
-			})
-		}
-	})
 }
 
 func (a *App) getSystemInstallDate() (int64, *model.AppError) {
