+++ app/saml.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/mattermost/mattermost-server/model"
 	"github.com/mattermost/mattermost-server/utils"
+	"github.com/mattermost/mattermost-server/utils/fileutils"
 )
 
 func (a *App) GetSamlMetadata() (string, *model.AppError) {
@@ -40,7 +41,7 @@ func WriteSamlFile(fileData *multipart.FileHeader) *model.AppError {
 	}
 	defer file.Close()
 
-	configDir, _ := utils.FindDir("config")
+	configDir, _ := fileutils.FindDir("config")
 	out, err := os.Create(filepath.Join(configDir, filename))
 	if err != nil {
 		return model.NewAppError("AddSamlCertificate", "api.admin.add_certificate.saving.app_error", nil, err.Error(), http.StatusInternalServerError)
@@ -112,7 +113,7 @@ func RemoveSamlFile(filename string) *model.AppError {
 		return model.NewAppError("AddSamlCertificate", "api.admin.remove_certificate.delete.app_error", nil, "", http.StatusBadRequest)
 	}
 
-	if err := os.Remove(utils.FindConfigFile(filename)); err != nil {
+	if err := os.Remove(fileutils.FindConfigFile(filename)); err != nil {
 		return model.NewAppError("removeCertificate", "api.admin.remove_certificate.delete.app_error", map[string]interface{}{"Filename": filename}, err.Error(), http.StatusInternalServerError)
 	}
