+++ store/sqlstore/upgrade.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/mattermost/mattermost-server/mlog"
 	"github.com/mattermost/mattermost-server/model"
+	"github.com/mattermost/mattermost-server/services/timezones"
 )
 
 const (
@@ -416,7 +417,7 @@ func UpgradeDatabaseToVersion49(sqlStore SqlStore) {
 
 	if shouldPerformUpgrade(sqlStore, VERSION_4_8_1, VERSION_4_9_0) {
 		sqlStore.CreateColumnIfNotExists("Teams", "LastTeamIconUpdate", "bigint", "bigint", "0")
-		defaultTimezone := model.DefaultUserTimezone()
+		defaultTimezone := timezones.DefaultUserTimezone()
 		defaultTimezoneValue, err := json.Marshal(defaultTimezone)
 		if err != nil {
 			mlog.Critical(fmt.Sprint(err))
