+++ utils/config.go
@@ -24,6 +24,7 @@ import (
 	"github.com/mattermost/mattermost-server/einterfaces"
 	"github.com/mattermost/mattermost-server/mlog"
 	"github.com/mattermost/mattermost-server/model"
+	"github.com/mattermost/mattermost-server/utils/fileutils"
 	"github.com/mattermost/mattermost-server/utils/jsonutils"
 )
 
@@ -33,13 +34,6 @@ const (
 )
 
 var (
-	commonBaseSearchPaths = []string{
-		".",
-		"..",
-		"../..",
-		"../../..",
-	}
-
 	termsOfServiceEnabledAndEmpty = model.NewAppError(
 		"Config.IsValid",
 		"model.config.is_valid.support.custom_terms_of_service_text.app_error",
@@ -48,87 +42,6 @@ var (
 		http.StatusBadRequest,
 	)
 )
-
-func FindPath(path string, baseSearchPaths []string, filter func(os.FileInfo) bool) string {
-	if filepath.IsAbs(path) {
-		if _, err := os.Stat(path); err == nil {
-			return path
-		}
-
-		return ""
-	}
-
-	searchPaths := []string{}
-	searchPaths = append(searchPaths, baseSearchPaths...)
-
-	// Additionally attempt to search relative to the location of the running binary.
-	var binaryDir string
-	if exe, err := os.Executable(); err == nil {
-		if exe, err = filepath.EvalSymlinks(exe); err == nil {
-			if exe, err = filepath.Abs(exe); err == nil {
-				binaryDir = filepath.Dir(exe)
-			}
-		}
-	}
-	if binaryDir != "" {
-		for _, baseSearchPath := range baseSearchPaths {
-			searchPaths = append(
-				searchPaths,
-				filepath.Join(binaryDir, baseSearchPath),
-			)
-		}
-	}
-
-	for _, parent := range searchPaths {
-		found, err := filepath.Abs(filepath.Join(parent, path))
-		if err != nil {
-			continue
-		} else if fileInfo, err := os.Stat(found); err == nil {
-			if filter != nil {
-				if filter(fileInfo) {
-					return found
-				}
-			} else {
-				return found
-			}
-		}
-	}
-
-	return ""
-}
-
-// FindConfigFile attempts to find an existing configuration file. fileName can be an absolute or
-// relative path or name such as "/opt/mattermost/config.json" or simply "config.json". An empty
-// string is returned if no configuration is found.
-func FindConfigFile(fileName string) (path string) {
-	found := FindFile(filepath.Join("config", fileName))
-	if found == "" {
-		found = FindPath(fileName, []string{"."}, nil)
-	}
-
-	return found
-}
-
-// FindFile looks for the given file in nearby ancestors relative to the current working
-// directory as well as the directory of the executable.
-func FindFile(path string) string {
-	return FindPath(path, commonBaseSearchPaths, func(fileInfo os.FileInfo) bool {
-		return !fileInfo.IsDir()
-	})
-}
-
-// FindDir looks for the given directory in nearby ancestors relative to the current working
-// directory as well as the directory of the executable, falling back to `./` if not found.
-func FindDir(dir string) (string, bool) {
-	found := FindPath(dir, commonBaseSearchPaths, func(fileInfo os.FileInfo) bool {
-		return fileInfo.IsDir()
-	})
-	if found == "" {
-		return "./", false
-	}
-
-	return found, true
-}
 
 func MloggerConfigFromLoggerConfig(s *model.LogSettings) *mlog.LoggerConfiguration {
 	return &mlog.LoggerConfiguration{
@@ -154,7 +67,7 @@ func EnableDebugLogForTest() {
 
 func GetLogFileLocation(fileLocation string) string {
 	if fileLocation == "" {
-		fileLocation, _ = FindDir("logs")
+		fileLocation, _ = fileutils.FindDir("logs")
 	}
 
 	return filepath.Join(fileLocation, LOG_FILENAME)
@@ -432,10 +345,10 @@ func ReadConfigFile(path string, allowEnvironmentOverrides bool) (*model.Config,
 // it will attempt to locate a default config file, and copy it to a file named fileName in the same
 // directory. In either case, the config file path is returned.
 func EnsureConfigFile(fileName string) (string, error) {
-	if configFile := FindConfigFile(fileName); configFile != "" {
+	if configFile := fileutils.FindConfigFile(fileName); configFile != "" {
 		return configFile, nil
 	}
-	if defaultPath := FindConfigFile("default.json"); defaultPath != "" {
+	if defaultPath := fileutils.FindConfigFile("default.json"); defaultPath != "" {
 		destPath := filepath.Join(filepath.Dir(defaultPath), fileName)
 		src, err := os.Open(defaultPath)
 		if err != nil {
